2024-10-23 - bollinger-bands-for-time-series-analysis-using-natural-gas-prices-with-python/requirements.txt
pandas
numpy
matplotlib
pyyaml
pytest
pandas-datareader
prophet
